requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
playwright==1.40.0
PyMuPDF==1.24.0
google-generativeai==0.3.2
tenacity==8.5.0
